+++ src/main/java/com/epam/reportportal/formatting/http/HttpRequestFormatter.java
@@ -227,9 +227,14 @@ public class HttpRequestFormatter implements HttpFormatter {
 			return this;
 		}
 
+		public Builder addCookie(Cookie cookie) {
+			cookies.add(cookie);
+			return this;
+		}
+
 		public Builder addCookie(String name, String value, String comment, String path, String domain, Long maxAge,
 				Boolean secured, Boolean httpOnly, Date expiryDate, Integer version, String sameSite) {
-			cookies.add(HttpFormatUtils.toCookie(name,
+			return addCookie(HttpFormatUtils.toCookie(name,
 					value,
 					comment,
 					path,
@@ -241,7 +246,6 @@ public class HttpRequestFormatter implements HttpFormatter {
 					version,
 					sameSite
 			));
-			return this;
 		}
 
 		public Builder addCookie(String name, String value) {